fix(): update slate and fix typings for slate content
