+++ ui/design/package.json
@@ -119,9 +119,9 @@
     "react-markdown": "7.0.1",
     "react-toastify": "7.0.4",
     "remark-gfm": "2.0.0",
-    "slate": "0.65.3",
-    "slate-history": "0.65.3",
-    "slate-react": "0.65.3",
+    "slate": "0.66.1",
+    "slate-history": "0.66.0",
+    "slate-react": "0.66.1",
     "ts-jest": "27.0.5"
   },
   "peerDependencies": {
