+++ ui/design/src/components/Editor/custom-types.ts
@@ -1,58 +1,4 @@
-import { BaseEditor } from 'slate';
-import { ReactEditor } from 'slate-react';
-import { HistoryEditor } from 'slate-history';
-
-export type ImageElement = {
-  type: 'image';
-  url: string;
-  size: {
-    width: string;
-    height: string;
-    naturalWidth: string;
-    naturalHeight: string;
-  };
-  children: EmptyText[];
-};
-
-export type LinkElement = { type: 'link'; url: string; children: EmptyText[] };
-
-export type MentionElement = {
-  type: 'mention';
-  name: string;
-  userName: string;
-  pubKey: string;
-  avatar: string;
-  ethAddress: string;
-  children: EmptyText[];
-};
-
-export type TagElement = {
-  type: 'tag';
-  name: string;
-  totalPosts: number;
-  children: EmptyText[];
-};
-type CustomElement =
-  | ImageElement
-  | LinkElement
-  | MentionElement
-  | TagElement
-  | { type: 'paragraph'; children: CustomText[] };
-
-export type CustomText = {
-  bold?: boolean;
-  italic?: boolean;
-  code?: boolean;
-  underlined?: boolean;
-  disabled?: boolean;
-  text: string;
-};
-
-export type EmptyText = {
-  text: string;
-};
-
-export type CustomEditor = BaseEditor & ReactEditor & HistoryEditor;
+import { CustomEditor, CustomElement, CustomText } from '@akashaproject/ui-awf-typings/lib/editor';
 
 declare module 'slate' {
   interface CustomTypes {
