+++ ui/typings/package.json
@@ -27,7 +27,9 @@
   "devDependencies": {
     "@akashaproject/sdk-typings": "*",
     "rxjs": "7.3.0",
-    "slate": "0.65.3",
+    "slate": "0.66.1",
+    "slate-react": "0.66.1",
+    "slate-history": "0.66.0",
     "i18next": "20.6.1",
     "single-spa": "5.9.3"
   },
